Small fixes
1. Added /productsPage
2. modified home and productsPage
3. fixed conflicting styles in global.css and home.css
