+++ src/app/global.css
@@ -28,7 +28,7 @@ html {
   background-color: lightpink;
   display: flex;
   flex-direction: column;
-  justify-content: flex-start;
+  justify-content: center;
   align-items: center;
   gap: 20px;
   overflow-y: auto;
