+++ src/app/home.css
@@ -1,15 +1,9 @@
-.Main {
-  width: 100%;
-  height: 100%;
-  background-color: lightpink;
+.home_main{
   display: flex;
   flex-direction: column;
-  justify-content: flex-start;
-  align-items: center;
-  gap: 20px;
-  overflow-y: auto;
+  justify-content: center;
+  gap: 50px;
 }
-
 .Main_Images {
   display: flex;
   justify-content: center;
@@ -18,7 +12,7 @@
 }
 
 .Main_Images img {
-  height: 150px;
+  height: 250px;
   width: auto;
 }
 
@@ -31,28 +25,11 @@
   color: gray;
 }
 
-.Product_List {
-  max-height: 250px;
-  overflow: auto;
-}
-.Product_List::-webkit-scrollbar {
-  width: 4px;
-}
-
-.Product_List::-webkit-scrollbar-thumb {
-  background: purple;
-  border-radius: 10px;
-}
-
 .Main_Grid::-webkit-scrollbar-track {
   background: pink;
 }
 
 @media (max-width: 1024px) {
-  .Product_List {
-    max-height: 200px;
-    margin-top: 40px;
-  }
 
   .Main_Images img {
     height: 100px;
